tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.size.attributes.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.size.attributes.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.attributes.default</h1>
<p class="desc">Default width/height when attributes are missing</p>


<script>
var t = async_test("Default width/height when attributes are missing");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(canvas.width, 100, "canvas.width", "100");
  _assertSame(canvas.height, 50, "canvas.height", "50");
  t.done();

});
</script>
